% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{calculateHDMWeights}
\alias{calculateHDMWeights}
\title{Calculate the normalized weights for each node and add them to the tree}
\usage{
calculateHDMWeights(tree, comboFrames)
}
\arguments{
\item{tree}{the tree representing the model to calculate}

\item{comboFrames}{a list of dataframes with comparative values corresponding
to the tree}
}
\description{
This function will take a defined model and corresponding set of
comboFrames and add normalized and weighted values to the tree
}
