% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{childrenOrNothing}
\alias{childrenOrNothing}
\title{If a node has children return them, otherwise return nothing}
\usage{
childrenOrNothing(currentNode)
}
\arguments{
\item{currentNode}{the node you want to check}
}
\description{
Sometimes you want to either the children of a node or nothing at all. This
function handles that for you.
}
