% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{disagreement.calculate}
\alias{disagreement.calculate}
\title{Pass in a list of experts and their final results, get back the disagreement}
\usage{
disagreement.calculate(expertsFinalResults)
}
\arguments{
\item{expertsFinalResults}{a matrix with the experts and the weights at the bottom
of the tree}
}
\description{
Pass in a list of experts and their final results, get back the disagreement
}
