% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{finalizeWeightsForNode}
\alias{finalizeWeightsForNode}
\title{Calculate the weighted value for each node}
\usage{
finalizeWeightsForNode(node)
}
\arguments{
\item{node}{the node to operate on}
}
\description{
As the final step in the calculation return the finalized weight of each node.
}
