% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{getExpertEvaluationComboFrames}
\alias{getExpertEvaluationComboFrames}
\title{Get the combo frames for an expert}
\usage{
getExpertEvaluationComboFrames(expertId, modelId, dataUri)
}
\arguments{
\item{expertId}{the expert you're looking for}

\item{modelId}{the model you're looking for}
}
\description{
Use this when you want to get the bitwise comparisons for an expert.
}
