% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{getExpertEvaluationRollup}
\alias{getExpertEvaluationRollup}
\title{Get all the expert results and roll them up for display}
\usage{
getExpertEvaluationRollup(experts, modelId, dataUri)
}
\arguments{
\item{experts}{collection of experts to get data for}

\item{modelId}{modelId that you're working with}

\item{dataUri}{URI of mongoDB to use}
}
\description{
Get all of the experts, calculate the summary stats across them,
return a nice matrix for display.
}
