% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{getExpertResultsAsTreeFromDb}
\alias{getExpertResultsAsTreeFromDb}
\title{Get an expert's evaluation in tree form}
\usage{
getExpertResultsAsTreeFromDb(modelId, expertId, dataUri)
}
\arguments{
\item{modelId}{the modelId for the mode you're looking for}

\item{expertId}{the expertId for the expert you're looking for}

\item{dataUri}{URI of mongoDB to use}
}
\description{
Given an expertId and modelId, get the evaluation for that expert
from the DB and return it in a tree context
}
