% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{getFullHDMModelFromDb}
\alias{getFullHDMModelFromDb}
\title{Get a full HDM model with alternatives and experts}
\usage{
getFullHDMModelFromDb(modelId, dataUri)
}
\arguments{
\item{modelId}{the modelId that you want to get}
}
\description{
Based on a modelId get an ad-hoc class that has everything we
need for an HDM based model
}
