% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{getLastElementInPath}
\alias{getLastElementInPath}
\title{Get the last element in a path string}
\usage{
getLastElementInPath(val)
}
\arguments{
\item{val}{the path to get the last element out of}
}
\description{
For some reason sometimes data.tree will return an entire path string
to a node instead of just the last element we need to rebuild the tree. This
helps fix that by taking the last element out of the path and returning it.
}
