% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{getModelAsTreeWithAlternativesFromDb}
\alias{getModelAsTreeWithAlternativesFromDb}
\title{Get the model in tree form with Alternatives as the bottom leaves}
\usage{
getModelAsTreeWithAlternativesFromDb(modelId, dataUri)
}
\arguments{
\item{modelId}{the modelId you're looking for}

\item{dataUri}{URI of mongoDB to use}
}
\description{
Get a model in tree form. If there are alternatives add them to the bottom
of the tree.
}
