% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.helper.r
\name{getNodeName}
\alias{getNodeName}
\title{Get the names of the nodes in a tree}
\usage{
getNodeName(node)
}
\arguments{
\item{node}{the node to get name of; meant to use recursively through Get method on tree}
}
\description{
Useful function used recursively to get a list of node names from a tree.
This somehow isn't default functionality from data.tree and is useful for UI
development.
}
