% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.helper.r
\name{getNodeNamesAtLevel}
\alias{getNodeNamesAtLevel}
\title{Return the names of nodes in a vector for the defined level}
\usage{
getNodeNamesAtLevel(tree, level)
}
\arguments{
\item{tree}{the tree you want to get data from}

\item{level}{the level of the tree you want to get}
}
\description{
Similar to getNodesAtLevel, but will only return the names. Also
useful for building operations.
}
