% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.helper.r
\name{getNodesAtLevel}
\alias{getNodesAtLevel}
\title{Get all the nodes at a specified level in the tree}
\usage{
getNodesAtLevel(tree, level)
}
\arguments{
\item{tree}{the tree you want to get data from}

\item{level}{the level of the tree you want to get}
}
\description{
Given a tree and a desired level, returns all nodes at the
specified level. This is useful for interface building purposes but
not a normal tree traversal operation.
}
