% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.helper.r
\name{getUniqueChildCombinations}
\alias{getUniqueChildCombinations}
\title{Get unique combinations for children of a node.
If there are no children compare alternatives}
\usage{
getUniqueChildCombinations(node, alternatives)
}
\arguments{
\item{node}{the node whos children you want to compare}

\item{alternatives}{OPTIONAL if you're at the end of the tree and
this node has no children, compare alternatives}
}
\description{
Get unique combinations for children of a node.
If there are no children compare alternatives
}
