% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.helper.r
\name{grid.unique}
\alias{grid.unique}
\title{Create a unique comparison matrix}
\usage{
grid.unique(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{first list to compare}

\item{y}{second list to compare}
}
\description{
This function will take in 2 lists and return the unique set of combinations
between them. This is necessary in HDM when you need to create the pairwise
comparisons based on children of a node in the hierarchy. This is only used
internally by getUniqueChildCombinations.
}
