% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.package.R
\docType{package}
\name{hdmr}
\alias{hdmr}
\alias{hdmr-package}
\title{hdm HDM (Hierarchical Decision Making) Modeling for R}
\description{
hdm HDM (Hierarchical Decision Making) Modeling for R
}
\examples{
library(hdpr)
TODO load up example data, run the stuff

}
