% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{inconsistency.calculate}
\alias{inconsistency.calculate}
\title{Calculate inconsistency}
\usage{
inconsistency.calculate(B)
}
\arguments{
\item{B}{the matrix after it's gone through the first 2 steps of the calculation}
}
\description{
Helps determine how reliable this expert is
}
