% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{loadAllModels}
\alias{loadAllModels}
\title{Utility to load all the models ids and names}
\usage{
loadAllModels(dataUri)
}
\arguments{
\item{dataUri}{URI of mongoDB to use}
}
\description{
Even though I use this, it's best not to expose it because you don't want
user's messing with each other's models. I think it's ok in an evironment of
total trust.
}
\details{
Load existing models by ID
}
