% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{loadHDMModel}
\alias{loadHDMModel}
\title{Load a model from the DB}
\usage{
loadHDMModel(modelId, dataUri)
}
\arguments{
\item{modelId}{the modelId that you're looking for}

\item{dataUri}{URI of mongoDB to use}
}
\description{
Internal function that makes the MongoDB Query
}
