% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{loadMyModelsFromDb}
\alias{loadMyModelsFromDb}
\title{Load modelIds and names for a particular user.}
\usage{
loadMyModelsFromDb(userEmail, pin, dataUri)
}
\arguments{
\item{userEmail}{the email of the user who build the models}

\item{pin}{acts like a password, only is not really secure}
}
\description{
Use this to get model identifiers  for models created by a particular user.
Note that I'm not
using real paswords, I'm not encrypting anything, and I'm not enforcing
much security. This is really just to keep users from messing with each
other's models on a shared tenant. If you want better security then you might
consider putting a proper authentication server in front of a user app or
something like that.
}
