% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{loadResults}
\alias{loadResults}
\title{Load expert evaluations from a MongoDB}
\usage{
loadResults(modelId, expertId, dataUri)
}
\arguments{
\item{modelId}{the ID of the model you want to load}

\item{expertId}{OPTIONAL get results from a specific expert}
}
\description{
Internal function to load up raw resuls from the DB
}
