% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{matrix.buildFromComboFrames}
\alias{matrix.buildFromComboFrames}
\title{Given a set of names and dataframes with evaluation data,
build out the first matrix to operate on}
\usage{
matrix.buildFromComboFrames(names, comboFrames)
}
\arguments{
\item{names}{the names of the elements being compared in a pairwise manner}

\item{comboFrames}{the frames to transform into a matrix}
}
\description{
This is only used internally as the first and second step in the calculation
}
