% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{matrix.calculate}
\alias{matrix.calculate}
\title{Given a populated matrix, divide the comparisons against
each other to get relative weights}
\usage{
matrix.calculate(A)
}
\arguments{
\item{A}{the matrix to operate on}
}
\description{
This is only used internally as the second part of the calculation
}
