% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{normalizeValueForNode}
\alias{normalizeValueForNode}
\title{Calculate weight for requested node}
\usage{
normalizeValueForNode(currentNode, comboFrames)
}
\arguments{
\item{currentNode}{the node to operate on}

\item{comboFrames}{the associated frames to use in the calculation}
}
\description{
Designed to be used recursively in the tree, this function will return
the normalized value for a node.
}
