% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{rebuildDataFrameForHDMTree}
\alias{rebuildDataFrameForHDMTree}
\title{Return a data.frame that can be converted to a tree
with your model based on data retrieved from somewhere else.}
\usage{
rebuildDataFrameForHDMTree(mod)
}
\arguments{
\item{model}{definition from database}
}
\value{
A \code{\link{data.tree}} containing the model
}
\description{
Return a data.frame that can be converted to a tree
with your model based on data retrieved from somewhere else.
}
