% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{saveHDMDataToMongoDb}
\alias{saveHDMDataToMongoDb}
\title{Save all data pertaining to the definition of a model to MongoDB}
\usage{
saveHDMDataToMongoDb(data, id, dataUri)
}
\arguments{
\item{data}{the full set of data to save in JSON format}

\item{id}{the ID of the model to update. If no ID is passed we will insert a record}

\item{dataUri}{URI of mongoDB to use}
}
\description{
Send in JSON data, this will save it to MongoDB. I love MongoDB because it's
web scale: https://www.youtube.com/watch?v=b2F-DItXtZs If you don't for some
reason then feel free to submit a PR to this repo for something else.
}
