% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{saveHdmEvaluationToDb}
\alias{saveHdmEvaluationToDb}
\title{Save an expert's evaluation to the DB}
\usage{
saveHdmEvaluationToDb(data, expertId, modelId, dataUri)
}
\arguments{
\item{data}{the results from the evaluation}

\item{expertId}{the ID of the expert to associate this evaluation with}

\item{modelId}{the ID of the model being evaluated}

\item{dataUri}{URI of mongoDB to use}
}
\description{
Once an expert evaluates all of their options this function will save them
so we can aggregate them and use them later.
}
