% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{trim}
\alias{trim}
\title{Returns string w/o leading or trailing whitespace}
\usage{
trim(x)
}
\arguments{
\item{x}{where x is the string you want to trim.}
}
\description{
This is a really useful function for building trees as whitespace is
valid in node names but messes up working with them later on.
}
