% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdm.calculator.r
\name{calculateExpertDisagreement}
\alias{calculateExpertDisagreement}
\title{Pass in a list of experts and their final results, get back the disagreement}
\usage{
calculateExpertDisagreement(expertsFinalResults)
}
\arguments{
\item{expertsFinalResults}{a matrix with the experts and the weights at the bottom
of the tree}
}
\description{
First get the standard deviation across the values for each expert
Second get the mean across the experts
}
