% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.functions.r
\name{getInconsistencyList}
\alias{getInconsistencyList}
\title{Get a list of data frames with experts and their inconsistency}
\usage{
getInconsistencyList(experts, modelId, dataUri)
}
\arguments{
\item{experts}{list of experts}

\item{modelId}{the model you're looking for}

\item{dataUri}{URI of mongoDB to use}
}
\description{
given a list of experts, a modelId, and a URI to get data from
get
}
